pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "migration-atlas"
version = "0.1.0"
description = "An NLP-powered knowledge graph and ML toolkit for U.S. immigration analysis"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [{ name = "Your Name", email = "you@example.com" }]
keywords = ["immigration", "nlp", "knowledge-graph", "machine-learning", "rag", "forecasting"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    # Core
    "pandas>=2.1",
    "numpy>=1.26",
    "pyarrow>=14.0",
    "pydantic>=2.5",
    "pydantic-settings>=2.1",
    "pyyaml>=6.0",

    # Graph
    "networkx>=3.2",
    "node2vec>=0.4.6",

    # NLP / ML
    "scikit-learn>=1.3",
    "transformers>=4.36",
    "torch>=2.1",
    "sentence-transformers>=2.2",
    "spacy>=3.7",
    "datasets>=2.16",
    "evaluate>=0.4",
    "accelerate>=0.25",

    # RAG
    "chromadb>=0.4",
    "langchain>=0.1",
    "langchain-community>=0.0.10",
    "pypdf>=3.17",

    # Forecasting
    "prophet>=1.1.5",
    "statsmodels>=0.14",

    # API + serving
    "fastapi>=0.108",
    "uvicorn[standard]>=0.25",
    "httpx>=0.26",

    # Utility
    "rich>=13.7",
    "typer>=0.9",
    "python-dotenv>=1.0",
    "tqdm>=4.66",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4",
    "pytest-cov>=4.1",
    "pytest-asyncio>=0.23",
    "black>=23.12",
    "ruff>=0.1.9",
    "mypy>=1.8",
    "ipykernel>=6.28",
    "jupyterlab>=4.0",
]
docs = [
    "mkdocs>=1.5",
    "mkdocs-material>=9.5",
    "mkdocstrings[python]>=0.24",
]
tracking = [
    "wandb>=0.16",
    "mlflow>=2.9",
]
neo4j = [
    "neo4j>=5.15",
]

[project.scripts]
migration-atlas = "migration_atlas.cli:app"

[project.urls]
Homepage      = "https://github.com/YOUR_USERNAME/migration-atlas"
Documentation = "https://YOUR_USERNAME.github.io/migration-atlas"
Repository    = "https://github.com/YOUR_USERNAME/migration-atlas"
Issues        = "https://github.com/YOUR_USERNAME/migration-atlas/issues"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
migration_atlas = ["data/*.json", "data/*.yaml"]

# ---------- Tooling ----------

[tool.black]
line-length = 100
target-version = ["py310"]

[tool.ruff]
line-length = 100
target-version = "py310"
select = ["E", "F", "W", "I", "N", "B", "UP", "SIM"]
ignore = ["E501"]  # handled by black

[tool.mypy]
python_version = "3.10"
strict = false
ignore_missing_imports = true
disallow_untyped_defs = false
warn_unused_ignores = true

[tool.pytest.ini_options]
minversion = "7.4"
addopts = "-ra --strict-markers --cov=migration_atlas --cov-report=term-missing"
testpaths = ["tests"]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "gpu: marks tests requiring GPU",
    "integration: integration tests requiring external services",
]
